Manage pages history refactor
